backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg[binary]>=3.1
pydantic>=2.5
pydantic-settings>=2.2
python-multipart>=0.0.9
python-jose[cryptography]>=3.3
openai>=1.40.0
pdfplumber>=0.10
requests>=2.31
